requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2